v0.2.5
Answers to questions:

#Why the [::-1]?
Because this reversed the sorted list. Note that a list is sorted from smallest to largest.

#THIS NEEDS A REMAKE TO CONSIDER AGE GROUPS
No need for that, if you print out the results dictionary, the age groups and events have been sorted out.

#if someone is disqualified it'll be their house which isnt an int
Well, you can tell that they have been disqualified if you see a "998" in an entry. Here, the entry is one attribute shorter.

#Callum why does that matter with the way that you're doing it? Also we can just ignore any disqualified students, they wont be getting any points.
Well, I wanted to include them just in case, but now that I think of it, I reckon we don't need to take them into account.

I'm also going to take out the names of participants as they are not needed either, right?